Edit: Created updateKeyStates, updateMouseButtonStates methods used for updating the state of input after the initial push/release states are recorded and need to be changed to None/Held on subsequent iterations of the game loop.
Added handling for control, alt & shift modifiers, and changed the KeyMap to use a new KeyData struct which contains all of that information for every key.

Changed InputState to PushState as it made more sense as the name.

Added uniformMousePosition which records what percentage across the screen the cursor currently is, allowing for generalized handling of position on the size of the window changing.

Minor changes - updated Vector2<float> to Vector2f to be consistent with GameObjectManager. Changed struct returns to be pass by reference.