Hacky implementation of the GameEngine in order to get a moving sprite with arrow key input.
Sprite is more or less implemented.
IRenderable was updated to add and remove from the RenderSystem whenever it is created or destroyed.
Camera was added to the register.
Camera is implemented.
View is implemented.
Camera object's transform is tied to the movement of the sfml view now.
Sprite's object's transform is tied to the movement of the sprite being rendered.